+++ src/main/java/br/com/techchallenge/TechChallengeApplication.java
@@ -2,7 +2,6 @@ package br.com.techchallenge;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
 public class TechChallengeApplication {
