+++ src/main/java/br/com/techchallenge/adapters/dto/RestauranteResponseDTO.java
@@ -1,4 +1,4 @@
-package br.com.techchallenge.adapters.controller.dto;
+package br.com.techchallenge.adapters.dto;
 
 public record RestauranteResponseDTO(
         String nome,
