+++ src/main/java/br/com/techchallenge/exception/InternalServerErrorException.java
@@ -1,0 +1,15 @@
+package br.com.techchallenge.exception;
+
+public class InternalServerErrorException extends Exception {
+
+    private String message;
+
+    public InternalServerErrorException(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+}
