+++ src/main/java/br/com/techchallenge/infra/service/RestauranteService.java
@@ -3,7 +3,7 @@ package br.com.techchallenge.infra.service;
 import br.com.techchallenge.infra.entity.RestauranteEntity;
 import br.com.techchallenge.infra.repository.DonoRestauranteRepository;
 import br.com.techchallenge.infra.repository.RestauranteRepository;
-import br.com.techchallenge.shared.exception.InternalServerErrorException;
+import br.com.techchallenge.shared.InternalServerErrorException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
